完成风力发电机 权重和故障率标准化

+++ src/main/java/com/njust/wows_knowledge_graph/domain/relationship/Have.java
@@ -22,7 +22,7 @@ public class Have implements Comparable<Have> {
     @TargetNode
     private Feature feature;
 
-    private int weigh;
+    private float weigh;
 
     @Override
     public int compareTo(Have o) {
